requirements.txt
langchain==0.3.27
apache-airflow==2.8.2
sqlalchemy==1.4.51
numpy==1.26.4
pandas==2.0.3
faiss-cpu==1.7.4
numba==0.59.1
psycopg2==2.9.10
langchain-community==0.3.29
langchain-openai==0.3.32
openai==1.102.0
flask==2.2.5
pyspark==3.5.1
shap==0.48.0
mlflow==3.3.2
python-telegram-bot==22.3
twilio==9.7.2
python-dotenv==1.1.1
requests==2.32.5
tiktoken==0.11.0
reportlab==4.4.3
matplotlib==3.10.6
kafka-python==2.0.2
llvmlite==0.42.0